SoftwareMarshall-main-update/requirements.txt
altgraph==0.17.2
backports.functools-lru-cache==1.6.4
cx-Freeze==5.1
cycler==0.10.0
dis3==0.1.3
drawnow==0.72.5
future==0.18.2
kiwisolver==1.1.0
matplotlib==2.2.5
numpy==1.16.6
pandas==0.24.2
pathlib==1.0.1
pefile==2019.4.18
Pillow==6.2.2
PyInstaller==3.6
pyparsing==2.4.7
PyPDF2==1.28.5
PyPubSub==3.3.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.0
reportlab==3.5.59
scipy==1.2.3
six==1.16.0
unicodecsv==0.14.1
wxPython==4.0.7.post2
